pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["udonself <ilyaayil2625@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.105.0"}
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
pydantic = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
